Updated mc3_p3/ template
This assignment used to combine parts of MC3-P2 and the current
assignment. Those parts have been split off, and the irrelevent
code moved or deleted. The assignments have been re-named, so if
you started work on MC3-P3 prior to this commit, you should pay
close attention to the changes.